+++ packages/analytics-js-common/src/constants/integrations/BingAds/constants.ts
@@ -1,6 +1,6 @@
+import { BINGADS_NAME as NAME, BINGADS_DISPLAY_NAME as DISPLAY_NAME } from '../Destinations';
+
 const DIR_NAME = 'BingAds';
-const NAME = 'BINGADS';
-const DISPLAY_NAME = 'Bing Ads';
 
 const DISPLAY_NAME_TO_DIR_NAME_MAP = { [DISPLAY_NAME]: DIR_NAME };
 const CNameMapping = {