+++ packages/analytics-js-common/src/constants/integrations/FacebookPixel/constants.ts
@@ -1,6 +1,9 @@
+import {
+  FACEBOOK_PIXEL_NAME as NAME,
+  FACEBOOK_PIXEL_DISPLAY_NAME as DISPLAY_NAME,
+} from '../Destinations';
+
 const DIR_NAME = 'FacebookPixel';
-const NAME = 'FACEBOOK_PIXEL';
-const DISPLAY_NAME = 'Facebook Pixel';
 
 const DISPLAY_NAME_TO_DIR_NAME_MAP = { [DISPLAY_NAME]: DIR_NAME };
 const CNameMapping = {