+++ packages/analytics-js-common/src/constants/integrations/Olark/constants.ts
@@ -1,6 +1,6 @@
+import { OLARK_NAME as NAME, OLARK_DISPLAY_NAME as DISPLAY_NAME } from '../Destinations';
+
 const DIR_NAME = 'Olark';
-const NAME = 'OLARK';
-const DISPLAY_NAME = 'Olark';
 
 const DISPLAY_NAME_TO_DIR_NAME_MAP = { [DISPLAY_NAME]: DIR_NAME };
 const CNameMapping = {