+++ packages/analytics-js-cookies/rollup.config.mjs
@@ -27,7 +27,6 @@ let outDirNpm = `${outDir}${variantSubfolder}`;
 const distName = 'rsaCookies';
 const modName = 'rsaCookies';
 
-
 export function getDefaultConfig(distName) {
   const version = process.env.VERSION || 'dev-snapshot';
 
@@ -35,9 +34,7 @@ export function getDefaultConfig(distName) {
     watch: {
       include: ['src/**'],
     },
-    external: [
-      ...Object.keys(pkg.peerDependencies || {})
-    ],
+    external: [...Object.keys(pkg.peerDependencies || {})],
     onwarn(warning, warn) {
       // Silence 'this' has been rewritten to 'undefined' warning
       // https://rollupjs.org/guide/en/#error-this-is-undefined
@@ -96,7 +93,7 @@ export function getDefaultConfig(distName) {
           open: true,
           gzipSize: true,
           brotliSize: true,
-        })
+        }),
     ],
   };
 }
@@ -155,11 +152,11 @@ const buildEntries = () => {
             {
               find: '@rudderstack/analytics-js-common',
               replacement: path.resolve('./dist/dts/packages/analytics-js-common/src'),
-            }
-          ]
+            },
+          ],
         }),
         dts(),
-        del({ hook: "buildEnd", targets: "./dist/dts" }),
+        del({ hook: 'buildEnd', targets: './dist/dts' }),
       ],
       output: [
         {
@@ -169,10 +166,10 @@ const buildEntries = () => {
         {
           file: `${outDir}/index.d.cts`,
           format: 'es',
-        }
-      ]
-    }
+        },
+      ],
+    },
   ];
-}
+};
 
 export default buildEntries();