+++ packages/analytics-js-integrations/src/integrations/Amplitude/browser.js
@@ -89,7 +89,7 @@ class Amplitude {
   }
 
   isLoaded() {
-    return Boolean(window.amplitude && window.amplitude.getDeviceId());
+    return Boolean(window.amplitude?.getDeviceId());
   }
 
   isReady() {