+++ packages/analytics-js-integrations/src/integrations/Sendinblue/browser.js
@@ -43,7 +43,6 @@ class Sendinblue {
   }
 
   isReady() {
-
     return this.isLoaded();
   }