+++ packages/analytics-js-cookies/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.4.18](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-cookies@0.4.17...@rudderstack/analytics-js-cookies@0.4.18) (2025-01-03)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.14.15`
+## [0.4.17](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-cookies@0.4.16...@rudderstack/analytics-js-cookies@0.4.17) (2024-12-17)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.14.14`
+
+### Bug Fixes
+
+* remove circular dependency in packages ([#1973](https://github.com/rudderlabs/rudder-sdk-js/issues/1973)) ([e525496](https://github.com/rudderlabs/rudder-sdk-js/commit/e5254964310c2c73baaf4d0655c3e4025c5e7d2b))
+
+## [0.4.16](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-cookies@0.4.15...@rudderstack/analytics-js-cookies@0.4.16) (2024-12-06)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.14.13`
 ## [0.4.15](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-cookies@0.4.14...@rudderstack/analytics-js-cookies@0.4.15) (2024-11-22)
 
 ### Dependency Updates